requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.7
django-cors-headers==3.14.0
sqlparse==0.4.3
djangorestframework==3.13.1
numpy==1.22.3
pandas==1.4.1
psycopg2-binary==2.9.3
pymongo==4.0.2
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
python-dotenv==0.19.2
